four_question/requirements.txt
pyogrio==0.5.1
pyproj==3.4.1
Rtree==0.9.7
shapely==2.0.1
geopandas==0.12.2
numpy==1.22.4
rasterio==1.3.6
patchify==0.2.3
opencv-python==4.7.0.72
matplotlib==3.5.1
tensorflow==2.10.0